+++ Changelog.md
@@ -3,6 +3,7 @@ RestrictEvents Changelog
 #### v1.0.2
 - Fixed patching CPU brand string with 8 core configurations
 - Fixed detecting CPU core count on some CPU models
+- Added single-core CPU brand string spoofing support
 
 #### v1.0.1
 - Disabled `MacPro7,1` RAM & PCI Expansion Slot UIs